Fixing double definition
Removed .h implementation in .cpp
